Add deploy cmd for wskdeploy which splitted from root cmd.
Move servicedeployer to utils, refactored code.

Delete several unused command file.
